Step 1K: Implement cart-service to product-service integration with Kafka event handling

+++ cart-service/src/main/java/com/example/cart_service/controller/CartController.java
@@ -28,7 +28,7 @@ public class CartController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Cart> getCartById(@PathVariable Integer id) {
+    public ResponseEntity<Cart> getCartById(@PathVariable Long id) {
         return cartService.getCartById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
